Tweak for circle (#23)
* fix DSK loading; IPF/CAPS also for circle
* ROM settings : replace map by array
* Remove unnecessary define; fix warnings
* FIX : Alpha channel set to "full" for sprites / border / background.
* Fix a test; Fix the missing LOG in FDC causing FDC to work incorrectly
* remove simple_filesystem; simple_regex is now for everybody
* add multilanguage to core

+++ CPCCoreEmu/FormatTypeRAW.cpp
@@ -6,7 +6,7 @@
 
 #include "FileAccess.h"
 
-#include "simple_stdio.h"
+#include <stdio.h>
 using namespace std;
 
 #define TXT_OUT_PROF _T("Duree load disk")
@@ -488,7 +488,7 @@ void FormatTypeRAW::HandleHWInfo(char* info, int size)
 int FormatTypeRAW::ComputeTrack(IDisk*& created_disk, int side, int track)
 {
    // Number of Index copy :
-   int indexNumber = index_list_.size();
+   unsigned int indexNumber = static_cast<unsigned int>(index_list_.size());
 
    if (indexNumber < 2)
       return -1;
@@ -879,7 +879,7 @@ FormatTypeRAW::OperationReturn FormatTypeRAW::GetFileListFromName(const char* fi
    }
    else
    {
-      // Otherwise, its a file : build a généric mask from this file
+      // Otherwise, its a file : build a gï¿½nï¿½ric mask from this file
       std::string full_path = GetFullPath(file_path);
       std::string filename = GetFileFromPath(file_path);
